LayoutTests/imported/w3c/web-platform-tests/html/dom/render-blocking/script-inserted-module-script-expected.txt

1

PASS Rendering is blocked before render-blocking resources are loaded
PASS Script-inserted render-blocking module script is evaluated
